Replace macros
