+++ README.md
@@ -215,10 +215,11 @@ Batch execution allows transactional execution of a set of commands
 ```typescript
 const commands = [
   ['CREATE TABLE TEST (id integer)'],
-  ['INSERT INTO TEST (id) VALUES (?)', [1]][
-    ('INSERT INTO TEST (id) VALUES (?)', [2])
-  ][('INSERT INTO TEST (id) VALUES (?)', [[3], [4], [5], [6]])],
+  ['INSERT INTO TEST (id) VALUES (?)', [1]],
+  [('INSERT INTO TEST (id) VALUES (?)', [2])],
+  [('INSERT INTO TEST (id) VALUES (?)', [[3], [4], [5], [6]])],
 ];
+
 const result = QuickSQLite.executeSqlBatch('myDatabase', commands);
 if (!result.status) {
   // result.status undefined or 0 === success
