+++ cpp/bindings.h
@@ -1,16 +1,10 @@
-/*
- * react-native-quick-sqlite.h
- *
- * Created by Oscar Franco on 2021/03/07
- * Copyright (c) 2021 Oscar Franco
- *
- * This code is licensed under the MIT license
- */
-
 #include <jsi/jsilib.h>
 #include <jsi/jsi.h>
 #include <ReactCommon/CallInvoker.h>
 
 using namespace facebook;
 
+namespace osp {
 void install(jsi::Runtime &rt, std::shared_ptr<react::CallInvoker> jsCallInvoker, const char *docPath);
+}
+
